refactored tests for quantify
